Separated the functions and codes of getting users for better file organization
